Refactor: 코드리뷰 반영, model 객체 분리
[StoreItem]을 의미하는 모델 생성.
데이터 요청 로직 분리 - Service 객체.
